First commit for the start of this engine system.
Currently I merged in the Core library and it will only be updated in tandem with this (Core_Deprecated holds where I ended with that on git).

I have set up the start of SDL integreation in terms of the standard game loop, event polling and some input event mapping as well we window and GL context creation.

Next will be some rendering systems. More detailed synopsis in ReadMe.