Added barebones render manager.
NOTE: this render manager should not be updated by the application manager but by the game manager. The application manager should just hold onto it as it is the top-most (and most likely only) renderer and should be passed into the game manager, but the game manager should not have control over it's lifetime.